Refactor assay type bar chart components
