+++ context/app/static/js/components/home-revision/AssayTypeBarChartContainer/index.js
@@ -1,0 +1,3 @@
+import AssayTypeBarChartContainer from './AssayTypeBarChartContainer';
+
+export default AssayTypeBarChartContainer;
