+++ context/app/static/js/components/home-revision/AssayTypeBarChartContainer/utils.js
