+++ context/app/static/js/pages/Home/HomeRevision.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import AssayTypeBarChart from 'js/components/home-revision/AssayTypeBarChart';
+import AssayTypeBarChartContainer from 'js/components/home-revision/AssayTypeBarChartContainer';
 import { TemporaryContainer } from './style';
 
 function HomeRevision() {
   return (
     <TemporaryContainer>
-      <AssayTypeBarChart />
+      <AssayTypeBarChartContainer />
     </TemporaryContainer>
   );
 }
